+++ crates/ludtwig-parser/CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT-VERSION
+- Fix HTML Tag name parser token collisions, which caused tags like `source` to not parse correctly
 
 # v0.4.0
 - Rewrite of the whole parser to make it lossless and implemented error recovery